Allowing buy command to use inventory instead of generating its own menu
