+++ commands/economy/buy.js
@@ -1,7 +1,7 @@
 const logger = require('../../util/logger');
 const shopItems = require('../../asset/shop/shopItems.json');
 const economy = require('../../features/economy/economy');
-
+const inv = require('../../features/economy/inventory');
 module.exports = {
   commands: 'buy',
   description: 'Buy Loot Box',
@@ -17,21 +17,6 @@ module.exports = {
     const userTag = message.author.tag;
     const userName = message.author.username;
     const userNickname = message.member.nickname;
-    let inventoryMenu = {
-      embed: {
-        color: 15277667,
-        title: `**${userNickname}'s Inventory**`,
-        thumbnail: {
-          url: 'https://cdn.discordapp.com/attachments/835471079303544834/845921169642618890/Screen_Shot_2021-05-23_at_3.06.38_PM.png',
-        },
-        fields: [],
-        footer: {
-          icon_url:
-            'https://cdn.discordapp.com/attachments/835471079303544834/846039122346377276/Screen_Shot_2021-05-23_at_10.57.00_PM.png',
-          text: 'placeholder',
-        },
-      },
-    };
 
     let invalidBox = [];
 
@@ -66,17 +51,10 @@ module.exports = {
         const totalBox = await economy.addBox(guildId, userId, addBox);
         // take away coins needed
         const pay = await economy.removeCoins(guildId, userId);
-        inventoryMenu.embed.footer.text = `Remaining Cock Coins: ${pay}`;
-        // prepare result for menu
-        const inventory = totalBox.replace(/[,]/g, '\n').replace(/[:]/g, ': ');
-        // send info to menu
-        inventoryMenu.embed.fields.push({
-          name: `**Boxes**`,
-          value: `\`\`\`${inventory}\`\`\``,
-          inline: true,
-        });
-        // send menu to channel
-        message.channel.send(inventoryMenu);
+
+        // call inventory menu
+        inv.inventory(message);
+
         // if not enough balance, prompt the user
       } else if (check === 0) {
         message.channel.send(`You don't have enough Cock Coins!`);
